+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ## 1.41.0 (Unreleased)
+
+IMPROVEMENTS:
+
+- Improve kvstore_backup_policy testcase use common method [GH-1044]
+
 ## 1.40.0 (April 20, 2019)
 
 FEATURES: